+++ lib/phpoffice/phpspreadsheet/src/PhpSpreadsheet/Helper/Sample.php
@@ -2,8 +2,11 @@
 
 namespace PhpOffice\PhpSpreadsheet\Helper;
 
+use PhpOffice\PhpSpreadsheet\Chart\Chart;
 use PhpOffice\PhpSpreadsheet\IOFactory;
+use PhpOffice\PhpSpreadsheet\Settings;
 use PhpOffice\PhpSpreadsheet\Spreadsheet;
+use PhpOffice\PhpSpreadsheet\Worksheet\Worksheet;
 use PhpOffice\PhpSpreadsheet\Writer\IWriter;
 use RecursiveDirectoryIterator;
 use RecursiveIteratorIterator;
@@ -11,6 +14,7 @@ use RecursiveRegexIterator;
 use ReflectionClass;
 use RegexIterator;
 use RuntimeException;
+use Throwable;
 
 /**
  * Helper class to be used in sample code.
@@ -76,6 +80,11 @@ class Sample
     {
         // Populate samples
         $baseDir = realpath(__DIR__ . '/../../../samples');
+        if ($baseDir === false) {
+            // @codeCoverageIgnoreStart
+            throw new RuntimeException('realpath returned false');
+            // @codeCoverageIgnoreEnd
+        }
         $directory = new RecursiveDirectoryIterator($baseDir);
         $iterator = new RecursiveIteratorIterator($directory);
         $regex = new RegexIterator($iterator, '/^.+\.php$/', RecursiveRegexIterator::GET_MATCH);
@@ -83,9 +92,14 @@ class Sample
         $files = [];
         foreach ($regex as $file) {
             $file = str_replace(str_replace('\\', '/', $baseDir) . '/', '', str_replace('\\', '/', $file[0]));
+            if (is_array($file)) {
+                // @codeCoverageIgnoreStart
+                throw new RuntimeException('str_replace returned array');
+                // @codeCoverageIgnoreEnd
+            }
             $info = pathinfo($file);
-            $category = str_replace('_', ' ', $info['dirname']);
-            $name = str_replace('_', ' ', preg_replace('/(|\.php)/', '', $info['filename']));
+            $category = str_replace('_', ' ', $info['dirname'] ?? '');
+            $name = str_replace('_', ' ', (string) preg_replace('/(|\.php)/', '', $info['filename']));
             if (!in_array($category, ['.', 'boostrap', 'templates'])) {
                 if (!isset($files[$category])) {
                     $files[$category] = [];
@@ -109,7 +123,7 @@ class Sample
      * @param string $filename
      * @param string[] $writers
      */
-    public function write(Spreadsheet $spreadsheet, $filename, array $writers = ['Xlsx', 'Xls']): void
+    public function write(Spreadsheet $spreadsheet, $filename, array $writers = ['Xlsx', 'Xls'], bool $withCharts = false, ?callable $writerCallback = null): void
     {
         // Set active sheet index to the first sheet, so Excel opens this as the first sheet
         $spreadsheet->setActiveSheetIndex(0);
@@ -118,9 +132,16 @@ class Sample
         foreach ($writers as $writerType) {
             $path = $this->getFilename($filename, mb_strtolower($writerType));
             $writer = IOFactory::createWriter($spreadsheet, $writerType);
+            $writer->setIncludeCharts($withCharts);
+            if ($writerCallback !== null) {
+                $writerCallback($writer);
+            }
             $callStartTime = microtime(true);
             $writer->save($path);
-            $this->logWrite($writer, $path, $callStartTime);
+            $this->logWrite($writer, $path, /** @scrutinizer ignore-type */ $callStartTime);
+            if ($this->isCli() === false) {
+                echo '<a href="/download.php?type=' . pathinfo($path, PATHINFO_EXTENSION) . '&name=' . basename($path) . '">Download ' . basename($path) . '</a><br />';
+            }
         }
 
         $this->logEndingNotes();
@@ -136,7 +157,7 @@ class Sample
      *
      * @return string
      */
-    private function getTemporaryFolder()
+    public function getTemporaryFolder()
     {
         $tempFolder = sys_get_temp_dir() . '/phpspreadsheet';
         if (!$this->isDirOrMkdir($tempFolder)) {
@@ -151,14 +172,12 @@ class Sample
      *
      * @param string $filename
      * @param string $extension
-     *
-     * @return string
      */
-    public function getFilename($filename, $extension = 'xlsx')
+    public function getFilename($filename, $extension = 'xlsx'): string
     {
         $originalExtension = pathinfo($filename, PATHINFO_EXTENSION);
 
-        return $this->getTemporaryFolder() . '/' . str_replace('.' . $originalExtension, '.' . $extension, basename($filename));
+        return $this->getTemporaryFolder() . '/' . str_replace('.' . /** @scrutinizer ignore-type */ $originalExtension, '.' . $extension, basename($filename));
     }
 
     /**
@@ -171,15 +190,69 @@ class Sample
     public function getTemporaryFilename($extension = 'xlsx')
     {
         $temporaryFilename = tempnam($this->getTemporaryFolder(), 'phpspreadsheet-');
+        if ($temporaryFilename === false) {
+            // @codeCoverageIgnoreStart
+            throw new RuntimeException('tempnam returned false');
+            // @codeCoverageIgnoreEnd
+        }
         unlink($temporaryFilename);
 
         return $temporaryFilename . '.' . $extension;
     }
 
-    public function log($message): void
+    public function log(string $message): void
     {
         $eol = $this->isCli() ? PHP_EOL : '<br />';
-        echo date('H:i:s ') . $message . $eol;
+        echo($this->isCli() ? date('H:i:s ') : '') . $message . $eol;
+    }
+
+    public function renderChart(Chart $chart, string $fileName): void
+    {
+        if ($this->isCli() === true) {
+            return;
+        }
+
+        Settings::setChartRenderer(\PhpOffice\PhpSpreadsheet\Chart\Renderer\MtJpGraphRenderer::class);
+
+        $fileName = $this->getFilename($fileName, 'png');
+
+        try {
+            $chart->render($fileName);
+            $this->log('Rendered image: ' . $fileName);
+            $imageData = file_get_contents($fileName);
+            if ($imageData !== false) {
+                echo '<div><img src="data:image/gif;base64,' . base64_encode($imageData) . '" /></div>';
+            }
+        } catch (Throwable $e) {
+            $this->log('Error rendering chart: ' . $e->getMessage() . PHP_EOL);
+        }
+    }
+
+    public function titles(string $category, string $functionName, ?string $description = null): void
+    {
+        $this->log(sprintf('%s Functions:', $category));
+        $description === null
+            ? $this->log(sprintf('Function: %s()', rtrim($functionName, '()')))
+            : $this->log(sprintf('Function: %s() - %s.', rtrim($functionName, '()'), rtrim($description, '.')));
+    }
+
+    public function displayGrid(array $matrix): void
+    {
+        $renderer = new TextGrid($matrix, $this->isCli());
+        echo $renderer->render();
+    }
+
+    public function logCalculationResult(
+        Worksheet $worksheet,
+        string $functionName,
+        string $formulaCell,
+        ?string $descriptionCell = null
+    ): void {
+        if ($descriptionCell !== null) {
+            $this->log($worksheet->getCell($descriptionCell)->getValue());
+        }
+        $this->log($worksheet->getCell($formulaCell)->getValue());
+        $this->log(sprintf('%s() Result is ', $functionName) . $worksheet->getCell($formulaCell)->getCalculatedValue());
     }
 
     /**
@@ -203,7 +276,10 @@ class Sample
         $callTime = $callEndTime - $callStartTime;
         $reflection = new ReflectionClass($writer);
         $format = $reflection->getShortName();
-        $message = "Write {$format} format to <code>{$path}</code>  in " . sprintf('%.4f', $callTime) . ' seconds';
+
+        $message = ($this->isCli() === true)
+            ? "Write {$format} format to {$path}  in " . sprintf('%.4f', $callTime) . ' seconds'
+            : "Write {$format} format to <code>{$path}</code>  in " . sprintf('%.4f', $callTime) . ' seconds';
 
         $this->log($message);
     }
